+++ app/src/main/java/labs/zero_one/roundo/PreferenceActivity.kt
@@ -1,0 +1,87 @@
+package labs.zero_one.roundo
+
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import android.view.MenuItem
+
+/**
+ * 主设置页面的 Activity
+ *
+ * 重写方法列表
+ * [onCreate]
+ * [onOptionsItemSelected]
+ *
+ * 注释参考
+ *
+ * @author lucka
+ * @since 0.1.1
+ */
+
+class PreferenceMainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_preference_main)
+
+        if (savedInstanceState == null) {
+            val preferenceFragment = PreferenceMainFragment()
+            supportFragmentManager
+                .beginTransaction()
+                .add(R.id.preferenceMainFrame, preferenceFragment)
+                .commit()
+        }
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
+        if (item != null) {
+            when (item.itemId) {
+                android.R.id.home -> {
+                    // Call onBackPress() when tap the back button on the toolbar instead of finish()
+                    onBackPressed()
+                    return true
+                }
+            }
+        }
+        return super.onOptionsItemSelected(item)
+    }
+
+}
+
+/**
+ * 关于页面的 Activity
+ *
+ * 重写方法列表
+ * [onCreate]
+ * [onOptionsItemSelected]
+ *
+ * @author lucka
+ * @since 0.1
+ */
+class PreferenceAboutActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_preference_about)
+
+        if (savedInstanceState == null) {
+            val preferenceAboutFragment = PreferenceAboutFragment()
+            supportFragmentManager
+                .beginTransaction()
+                .add(R.id.preferenceAboutFrame, preferenceAboutFragment)
+                .commit()
+        }
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
+        if (item != null) {
+            when (item.itemId) {
+                android.R.id.home -> {
+                    onBackPressed()
+                    return true
+                }
+            }
+        }
+        return super.onOptionsItemSelected(item)
+    }
+
+}