+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 <h1 align=center>更新日志</h1>
 
 ```markdown
+## [0.1.1] - 2018-08-16
+- 0.1(10) -> 0.1.1(34)
+
+### Finished
+- App Icon
+- App frame including preference
+
+### Changed
+- Package name: lab.zero_one -> labs.zero_one
+```
+
+```markdown
 ## [0.1] - 2018-08-15
 - 0.1(10)
 - Initial version